saper/requirements.txt
# pip install -r requirements.txt
# pip freeze > requirements.txt

PyQt5==5.15.2
PyQt5-sip==12.8.1
PySide2==5.15.2
PyYAML==5.3.1
shiboken2==5.15.2
